// packages/api/redemptions/src/constants/environment.ts
export enum EnvironmentKeys {
  // Database
  REDEMPTIONS_DATABASE_HOST = 'REDEMPTIONS_DATABASE_HOST',
  REDEMPTIONS_DATABASE_NAME = 'REDEMPTIONS_DATABASE_NAME',
  REDEMPTIONS_DATABASE_PASSWORD = 'REDEMPTIONS_DATABASE_PASSWORD',
  REDEMPTIONS_DATABASE_PORT = 'REDEMPTIONS_DATABASE_PORT',
  REDEMPTIONS_DATABASE_SEED_METHOD = 'REDEMPTIONS_DATABASE_SEED_METHOD',
  REDEMPTIONS_DATABASE_TYPE = 'REDEMPTIONS_DATABASE_TYPE',
  REDEMPTIONS_DATABASE_USER = 'REDEMPTIONS_DATABASE_USER',
  REDEMPTIONS_DANGEROUSLY_ALLOW_DATABASE_SETUP_STRATEGY = 'REDEMPTIONS_DANGEROUSLY_ALLOW_DATABASE_SETUP_STRATEGY',

  // Integrations
  CODES_REDEEMED_ENVIRONMENT = 'CODES_REDEEMED_ENVIRONMENT',
  CODES_REDEEMED_HOST = 'CODES_REDEEMED_HOST',
  CODE_REDEEMED_PATH = 'CODE_REDEEMED_PATH',
  CODE_ASSIGNED_REDEEMED_PATH = 'CODE_ASSIGNED_REDEEMED_PATH',
}